requirements.txt
toml
git+https://github.com/quantumai-asean/model-card-toolkit.git
git+https://github.com/quantumai-asean/streamlit-pydantic.git
streamlit>=1.31
psycopg[binary]
pycountry
fairlearn
google-generativeai 

